update app with to better support updates from master
